+++ src/lib.rs
@@ -5,6 +5,12 @@
 //! * [API Docs](https://docs.rs/tetra)
 //! * [FAQ](https://tetra.seventeencups.net/FAQ)
 //!
+//! ## Status
+//!
+//! Tetra is no longer being actively developed, as of January 2022. Bug fixes and dependency updates may still happen from time to time, but no new features are planned. Feature PRs may be accepted, as long as they do not come with a large maintainence burden - please open an issue/discussion thread if you're thinking about making any large changes!
+//!
+//! For more information, see [this blog post](https://www.seventeencups.net/posts/three-years-of-tetra/).
+//!
 //! ## Features
 //!
 //! * XNA/MonoGame-inspired API
@@ -21,7 +27,7 @@
 //! To add Tetra to your project, add the following line to your `Cargo.toml` file:
 //!
 //! ```toml
-//! tetra = "0.7"
+//! tetra = "0.8"
 //! ```
 //!
 //! You will also need to install the SDL2 native libraries - full details are provided in the [documentation](https://tetra.seventeencups.net/installation).